started integrating cell class and cell function
until now, the cell class that I created was pretty much unused, and the cell's of the board that get filled were entirely blitted and kept track of through a limited function. The cell class is now partially active and tracks Ids of each cell being created. Also have a preliminary version of the generate cells class function that will replace the current, global-scoped version.