important addition to Cell class, more integration
The merge is almost complete. Added x and y position as attributes the Cell class, as well as a class method to blit its value inside of its square. This let me simplify code even in the old blit and generate method for GridManager. Now the grid_rows in the main grid can directly accept cells, rather than dictionaries