created post_game(), enabled game_mode switch, etc
post_game() screen is a very basic placeholder for what will display the outcome of the game, replay saving options, and stats. Game mode switching is now enabled in the main_menu(), however larger boards will likely not work, since they don't fit on the screen. Changed GameHandler.game_is_active attribute to .game_status and instead of a boolean, it takes a string that describes the ongoing state of the game or the outcome.