bug fix, sound menu return button change
A bug was fixed where the tie_check would run after a win check determined a winner if the grid was full. A couple of text mistakes were fixed. The return button now sends the player back to the main menu if a game isn't ongoing and back into the game if it is. Created the "pregame" game_status under GameTracker to achieve this.