major replay player progress, reset implemented
I finally got the time gate logic correct. So now both moves in a turn
play out in sequence upon a button press with a 1 second gap between
them. Can now reset the replays with a button press (replaced ff with
reset). Maybe fixed a bug that ties what should be a win on the last
move in a tictactoe grid.