CPU logic changes, advancement
Made building off a single cell a new, lower class of move called
"constructive". Optimal is now reserved for moves that build on 2 or
more cells, and currently they are only implemented for descending
diagonals. Also set difficulty to 0 for testing's sake, and switched
around the signs for success/fail tests for readability.